Add files via upload
This is the game mechanic for placing blocks that is a component for the player. It can place and destroy blocks.

Things needed on your side:
Block prefab with collider
ShootingPoint empty Game Object that is child of camera and place directly in front and a bit above the camera
Tags called 'Block' and 'Ground'

Find out how your player can look downwards more freely because this hinders block placing, preventing nerd poling.